rename: store 이름 placeStore로 변경 및 import 경로 수정
